+++ crates/objc2/src/__macro_helpers/mod.rs
@@ -1,6 +1,6 @@
-pub use core::borrow::{Borrow, BorrowMut};
+pub use core::borrow::Borrow;
 pub use core::cell::UnsafeCell;
-pub use core::convert::{AsMut, AsRef};
+pub use core::convert::AsRef;
 pub use core::marker::{PhantomData, Sized};
 pub use core::mem::{size_of, ManuallyDrop, MaybeUninit};
 pub use core::ops::{Deref, DerefMut};