+++ crates/objc2/CHANGELOG.md
@@ -49,6 +49,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * **BREAKING**: Disallow `&mut` message receivers (except in the special case
   when the object is `AnyObject`, for better backwards compatibility with
   `objc`).
+* **BREAKING**: Removed `AsMut` and `BorrowMut` implementations in
+  `extern_class!` and `declare_class!`.
 
 ### Fixed
 * Remove an incorrect assertion when adding protocols to classes in an unexpected